+++ webserver/merchant_handlers_integration_test.go
@@ -127,7 +127,7 @@ func TestMerchantHandlersAPI(t *testing.T) {
 			testConfig.TransactionIDGenerator = &mock.StringIDGenerator{MockID: mockTransactionID}
 
 			Convey("When card has sufficient funds", func() {
-				transactionAmount := int32(50)
+				transactionAmount := 50
 
 				requestBody := authRequestBody(mockCardNumber, mockMerchantID, transactionAmount)
 				output := simulatePost(testConfig, "/v1/merchant/authorize-transaction", requestBody)
@@ -161,7 +161,7 @@ func TestMerchantHandlersAPI(t *testing.T) {
 			})
 
 			Convey("When card has insufficient funds", func() {
-				largeAmount := int32(5000)
+				largeAmount := 5000
 				requestBody := authRequestBody(mockCardNumber, mockMerchantID, largeAmount)
 				output := simulatePost(testConfig, "/v1/merchant/authorize-transaction", requestBody)
 
@@ -206,7 +206,7 @@ func TestMerchantHandlersAPI(t *testing.T) {
 				Convey("Returns 400", func() {
 					So(output.Code, ShouldEqual, 400)
 				})
-				
+
 				Convey("Returns error message", func() {
 					bodyAsString := output.Body.String()
 					So(bodyAsString, ShouldEqual, `{"error":"can not over-capture"}`)
@@ -255,10 +255,78 @@ func TestMerchantHandlersAPI(t *testing.T) {
 				})
 			})
 		})
+
+		// Convey("POST /v1/merchant/reverse-transaction", func() {
+		// 	mockTransactionID := "TX88888"
+		// 	authorizedTransaction := model.Transaction{
+		// 		ID:         mockTransactionID,
+		// 		CardNumber: mockCardNumber,
+		// 		Authorized: 100,
+		// 	}
+		// 	db.StoreMerchant(model.Merchant{
+		// 		ID:           mockMerchantID,
+		// 		Transactions: []model.Transaction{authorizedTransaction},
+		// 	})
+
+		// 	Convey("When reverse is possible", func() {
+		// 		requestBody := reverseRequestBody(mockMerchantID, mockTransactionID, 170)
+		// 		output := simulatePost(testConfig, "/v1/merchant/reverse-transaction", requestBody)
+		// 		Convey("Returns 400", func() {
+		// 			So(output.Code, ShouldEqual, 400)
+		// 		})
+
+		// 		Convey("Returns error message", func() {
+		// 			bodyAsString := output.Body.String()
+		// 			So(bodyAsString, ShouldEqual, `{"error":"can not over-reverse"}`)
+		// 		})
+
+		// 		Convey("Leaves DB unaffected", func() {
+		// 			merchant, _ := db.GetMerchant(mockMerchantID)
+		// 			So(merchant.Transactions, ShouldResemble, []model.Transaction{
+		// 				model.Transaction{
+		// 					ID:         mockTransactionID,
+		// 					CardNumber: mockCardNumber,
+		// 					Authorized: 100,
+		// 					Reversed:   0,
+		// 				}})
+		// 		})
+		// 	})
+
+		// 	Convey("When reverse is not possible", func() {
+		// 		requestBody := reverseRequestBody(mockMerchantID, mockTransactionID, 70)
+		// 		output := simulatePost(testConfig, "/v1/merchant/reverse-transaction", requestBody)
+		// 		Convey("Returns 200", func() {
+		// 			So(output.Code, ShouldEqual, 200)
+		// 		})
+
+		// 		Convey("Moves funds from Authorized Transactions to reversed transactions", func() {
+		// 			merchant, _ := db.GetMerchant(mockMerchantID)
+		// 			So(merchant.Transactions, ShouldResemble, []model.Transaction{
+		// 				model.Transaction{
+		// 					ID:         mockTransactionID,
+		// 					CardNumber: mockCardNumber,
+		// 					Authorized: 30,
+		// 					Reversed:   70,
+		// 				}})
+		// 		})
+		// 	})
+
+		// 	Convey("When request is badly formed", func() {
+		// 		output := simulatePost(testConfig, "/v1/merchant/reverse-transaction", strings.NewReader("bad data"))
+		// 		Convey("Returns 400", func() {
+		// 			So(output.Code, ShouldEqual, 400)
+		// 		})
+
+		// 		Convey("Returns bad JSON response", func() {
+		// 			bodyAsString := output.Body.String()
+		// 			So(bodyAsString, ShouldEqual, `{"error":"bad JSON format"}`)
+		// 		})
+		// 	})
+		// })
 	})
 }
 
-func authRequestBody(cardNumber int64, merchantID string, transactionAmount int32) io.Reader {
+func authRequestBody(cardNumber int64, merchantID string, transactionAmount int) io.Reader {
 	return strings.NewReader(fmt.Sprintf(
 		`{
 			"card_number": %d,
@@ -270,7 +338,7 @@ func authRequestBody(cardNumber int64, merchantID string, transactionAmount int3
 		transactionAmount,
 	))
 }
-func captureRequestBody(merchantID string, tranasactionID string, transactionAmount int32) io.Reader {
+func captureRequestBody(merchantID string, tranasactionID string, transactionAmount int) io.Reader {
 	return strings.NewReader(fmt.Sprintf(
 		`{
 			"merchant_id": "%s",
